requirements.txt
Flask
python-telegram-bot==20.3
pandas
openpyxl
